refactor(agent-runtime): flatten a2a.client sub-package into engine.a2a
Merge A2aCardAutoConfiguration into A2aClientAutoConfiguration and eliminate
the a2a.client sub-package:

- Move RemoteAgentCardCache, RemoteAgentInvocationService,
  A2aRemoteAgentOutboundAdapter, RemoteAgentProperties from a2a.client to a2a
- Merge A2aCardAutoConfiguration (AgentCard bean + AgentCardProperties) into
  A2aClientAutoConfiguration as the unconditional outer bean; remote support
  beans remain conditional under the RemoteConfiguration inner class
- Remove A2aCardAutoConfiguration class
- Update AutoConfiguration.imports to single A2A entry
- Update cross-package imports in openjiuwen adapter and all test files
