+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/engine/a2a/A2aParentTaskProjector.java
@@ -2,7 +2,6 @@ package com.huawei.ascend.runtime.engine.a2a;
 
 import com.huawei.ascend.runtime.common.RuntimeIdentity;
 import com.huawei.ascend.runtime.engine.AgentExecutionContext;
-import com.huawei.ascend.runtime.engine.a2a.client.RemoteAgentInvocationService;
 import com.huawei.ascend.runtime.engine.spi.AgentExecutionResult;
 import java.util.HashMap;
 import java.util.List;
