+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/engine/a2a/A2aRemoteAgentOutboundAdapter.java
@@ -1,4 +1,4 @@
-package com.huawei.ascend.runtime.engine.a2a.client;
+package com.huawei.ascend.runtime.engine.a2a;
 
 import java.time.Duration;
 import java.util.ArrayList;
