+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/engine/a2a/RemoteAgentCardCache.java
@@ -1,4 +1,4 @@
-package com.huawei.ascend.runtime.engine.a2a.client;
+package com.huawei.ascend.runtime.engine.a2a;
 
 import java.net.URI;
 import java.util.ArrayList;
