+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/engine/a2a/RemoteAgentProperties.java
@@ -1,4 +1,4 @@
-package com.huawei.ascend.runtime.engine.a2a.client;
+package com.huawei.ascend.runtime.engine.a2a;
 
 import java.util.List;
 import org.springframework.boot.context.properties.ConfigurationProperties;
