+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/engine/a2a/RemoteSupport.java
@@ -1,6 +1,5 @@
 package com.huawei.ascend.runtime.engine.a2a;
 
-import com.huawei.ascend.runtime.engine.a2a.client.RemoteAgentInvocationService;
 import java.util.Objects;
 
 /**
