+++ agent-runtime/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,3 +1,2 @@
 com.huawei.ascend.runtime.boot.RuntimeAutoConfiguration
-com.huawei.ascend.runtime.engine.a2a.A2aCardAutoConfiguration
-com.huawei.ascend.runtime.engine.a2a.client.A2aClientAutoConfiguration
+com.huawei.ascend.runtime.engine.a2a.A2aClientAutoConfiguration
