+++ agent-runtime/src/test/java/com/huawei/ascend/runtime/engine/a2a/A2aRemoteAgentOutboundAdapterTest.java
@@ -2,8 +2,6 @@ package com.huawei.ascend.runtime.engine.a2a;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.huawei.ascend.runtime.engine.a2a.client.A2aRemoteAgentOutboundAdapter;
-import com.huawei.ascend.runtime.engine.a2a.client.RemoteAgentInvocationService;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
